Import from: prelink-cross-test.patch
Source: Wind River Systems

The purpose of this patch is to allow for the testsuite to be run
using the cross prelink.  This disables the run-time test, but will
keep all of the prelink testing.  To have full test coverage you will
want to execute these tests on the target as well.