Import from: prelink-endian-swap.patch
Source: Wind River Systems

When using elfutils/libelf in prelink, it attempts to be helpful
and do endian translations of most types.  The ELF_T_BYTE type is not
translated, and is used for the majority of the RELOCs, but not all
of them.

The patch only does internal endian swapping when the type is
ELF_T_BYTE.