Fix MSB shift in prelink adjusting debug info
This patch fixes an incorrect shift by 58 instead of 56 when computing the
most significant byte when prelink adjusts debug info.  Please commit if
OK.

This problem manifests itself as test failures for SPARC64 but not other
platforms I've tested, probably because the mmap_base setting for that
platform has a nonzero MSB and most others don't.

2009-08-06  Joseph Myers  <joseph@codesourcery.com>

	* src/dwarf2.c (dwarf2_write_le64, dwarf2_write_be64): Shift by 56
	not 58 for MSB.