Fix MIPS debug info prelinking
Hi all,

Prelink currently fails to relocate the debug info in MIPS libraries
because they use a custom section type.

Here is an example section from PPC:

   [29] .debug_frame      PROGBITS        00000000 0021b0 00002c 00
  0   0  4

The type is "PROGBITS", which is what prelink likes, so it works correctly.

Now, here is an example from an equivalent MIPS binary:

   [28] .debug_frame      MIPS_DWARF      00000000 000ee8 00002c 00
  0   0  4

Here the type is "MIPS_DWARF", a type not handled by prelink.

Please accept the attached patch to fix this.

Thanks

Andrew

2010-05-06  Andrew Stubbs  <ams@codesourcery.com>

	* src/dso.c (adjust_dso): Support MIPS debug sections.
	* src/prelink.h (SHT_MIPS_DWARF): New define.