Handle gap between .rel.dyn and .rel.plt section
This patch makes prelink handle a case when .rel.dyn and .rel.plt
sections are adjacent to each other in the sense that there is no
section in between, but have a zero-filled gap between them -- e.g., due
to a linker optimization.  Such a gap can be introduced at least on
MIPS.

The linker possibly leaving a gap between sections seems reasonable, and
attached is the patch for prelink to handle it.  I looked into binutils
to try fix ld from generating the gap in the first place, but relocation
processing is convoluted, and changing it is likely to introduce new
bugs.

2012-03-19  Maxim Kuvyrkov  <maxim@codesourcery.com>

        * src/reloc.c (find_reloc_sections): Don't error due to a gap between
        sections.

Commiter note:
While fixing binutils to not insert the gap is likely the correct
answer, I agree with Maxim that making the prelinker more tolerant of
this condition is a good thing.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>