m4/libelf.m4: Add unistd.h to test file
the libelf configure test in libelf.m4 uses functions
declared in <unistd.h>, so should include that header.

(I don't know the origin of this patch; it appears in our sources as part
a bulk commit of existing prelink changes in 2006.)

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>

2012-05-28  Joseph Myers  <joseph@codesourcery.com>
        * m4/libelf.m4: Add unistd.h

MH: I've verified the origin of this patch is 2006 or before, somehow
this was missed in the original cross-prelink merges.