Prelink patch for address range handling
One of the differences causing the testsuite failures I described resulted
from failure to adjust a DW_AT_high_pc address in debug info.  As
described in the DWARF4 specification, "The high PC value may be beyond
the last valid instruction in the executable." - it describes "the address
of the first location past the last instruction associated with the
entity".  So prelink needs to handle it pointing to the address just after
the end of the valid text addresses, which this patch does by subtracting
1.

The remaining differences appear to result from location list handling.
If I comment out the addr_to_sec call in adjust_location_list (as well as
applying the first two patches), that fixes all the testsuite failures,
but that's certainly not a correct change; I think all the addr_to_sec
calls from there should be passing and a bug elsewhere is causing some to
fail.

----

2012-09-08  Joseph Myers  <joseph@codesourcery.com>
        * src/dwarf2.c: Fix address range handling related to DW_AT_high_pc

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>