Prelink patch for address range handling
This patch addresses the problem that the same data address in
.debug_loc can be referenced more than once in .debug_info, resulting in
it being adjusted more than once - there is no guarantee about whether
the adjusted address will be within the existing range of data
addresses, and so pass addr_to_sec and be wrongly adjusted a second
time, or not.  This is fixed by using a hash to avoid adjusting the same
bit of debug info more than once.

----

2012-09-08  Joseph Myers  <joseph@codesourcery.com>
        * src/dwarf2.c: Fix address range handling when dup data addr

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>