Fix for "Could not find variable copy reloc is against" for x86_64
For certain random seeds, the error "Could not find variable copy reloc
is against" on x86_64.  The problem was use of host size_t in prelink-rtld
when target size_t was needed, resulting in certain calculations of
addresses in prelink being off by 0x100000000 forthe 32-bit host, 64-bit
target combination.  This patch arranges for the relevant code to use
either uint32_t or uint64_t depending on the ELF file in question (since
I'm not confident it would be safe to use a 64-bit type unconditionally).

------

2012-09-12  Joseph Myers  <joseph@codesourcery.com>
        * src/rtld/dl-lookup.c: Fix variable copy reloc when host/target
          byte size is different

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>