Fix prelink section ordering on subsequent prelink
Some linkers create a binary which has a section header table that is
not in section header offset order.  However, as noted in check_dso(),
several routines in prelink and in libelf-0.7.0 too rely on sh_offsets
monotonically increasing, and if that fails then prelink quits.  But
the check is only on dso's, not binaries.  For binaries, fdopen_dso()
reorders the section headers to section header offset order prior to
work on the section data, so that the new prelinked binary is written
in section order.
However, on subsequent prelinks of a prelinked binary, the
prelink_exec() works on the undone section, not the fdopen_dso()
reordered sections, and may construct a section header table with
section orders different than the first prelink of the binary.
The function below tests for that case and resets the section header
order to the offset order by enforcing the binutils linker change:
'The net effect is .shstrab section is now placed after .symtab and
.strtab sections'. See binutils-gdb.git commit 3e19fb8f990e4c.

Signed-off-by: Dave Lerner <dave.lerner@windriver.com>

2016-06-03   Dave Lerner <dave.lerner@windriver.com>
        * src/exec.c: New check_reorder_needed, section_reorder_fix functions

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>