Fix x32 layout1 test.
The layout1 test case currently fails for x32 because there is
insufficient address space for all the libraries.

There's two ways to fix this:

1. Increase mmap_end such that there's more space.

2. Reduce max_page_size so that less space is wasted.

I'm not sure what the implications of option 2 would be, so the attached
patch implements option 1.

This still limits the number of possible libraries somewhat severely,
but it's more than enough to pass the layout1 test.

Andrew

2016-12-16  Andrew Stubbs  <ams@codesourcery.com>

	* src/arch-x86_64.c (PL_ARCH(x32)): Set mmap_end to 0x60000000.